Split Main to helper file NodeMain in runtime folder
